// src/app/models/messages/ORMMessageDetail.ts
export class ORMMessageDetail {
    message_detail_id: String;
    cc_reciever: Boolean;
    recieve_date: String;
    readed: Boolean;
    mail_status: String;
    messages_id: String;
    reciever_id: String;
    location: String;
    deleted: Boolean;
    created_user: String;
    client_date_created: String;
    modified_user: String;
    client_date_modified: String;
    date_created: String;
    date_modified: String;
}